Move sct investigate functionality to separate file
Storing functions for sct investigate in utils.py cause
error with import(cross import of same modules).
