+++ sct.py
@@ -7,9 +7,9 @@ from prettytable import PrettyTable
 
 from sdcm.results_analyze import PerformanceResultsAnalyzer
 from sdcm.sct_config import SCTConfiguration
+from sdcm.sct_investigate import list_logs_by_test_id, restore_monitoring_stack
 from sdcm.utils import (list_instances_aws, list_instances_gce, clean_cloud_instances,
-                        aws_regions, get_scylla_ami_versions, get_s3_scylla_repos_mapping,
-                        list_logs_by_test_id, restore_monitoring_stack)
+                        aws_regions, get_scylla_ami_versions, get_s3_scylla_repos_mapping)
 
 click_completion.init()
